Extracted base web handler to separate module
